+++ test/hkimjp/wil2/core_test.clj
@@ -1,0 +1,8 @@
+(ns hkimjp.wil2.core-test
+  (:require [clojure.test :refer [deftest testing is]]
+            [hkimjp.wil2.core :as core]))
+
+(deftest a-test
+  (testing "I success."
+    (is (= 0 0))))
+